+++ pkg/apply.js
@@ -117,12 +117,12 @@ module.exports = async function ({ overwrite_version, friendly_errors } = {}) {
         `file access forbidden - ${
           process.platform === 'win32'
             ? 'make sure your user has elevated permissions.'
-            : `try running "sudo chmod a+wr -R ${err.path.replace(
+            : `try running "sudo chmod -R a+wr ${err.path.replace(
                 'Notion.app',
                 'Notion'
               )}" ${
                 err.dest
-                  ? `and "sudo chmod a+wr -R ${err.dest.replace(
+                  ? `and "sudo chmod -R a+wr ${err.dest.replace(
                       'Notion.app',
                       'Notion'
                     )}"`